Adding print funcs 3 and print u function
