+++ 2-print_funcs.c
@@ -25,10 +25,7 @@ int printb(va_list b, int count)
 */
 int printu(va_list u, int count)
 {
-	(void)u;
-	printf("printb");
-	print_number(count, count);
-	printf("%d\n", count);
+	count = print_un_number(va_arg(u, unsigned int), count);
 	return (count);
 }
 
@@ -74,51 +71,6 @@ int printX(va_list X, int count)
 	(void)X;
 	(void)count;
 	printf("printX");
-	return (count);
-}
-
-/**
-* printS - output a string with non-printable characters to standard output
-* @S: pointer to argument of type string literal
-* @count: existing total number of characters printed
-*
-* Return: cumulative total number of characters printed
-*/
-int printS(va_list S, int count)
-{
-	(void)S;
-	(void)count;
-	printf("printS");
-	return (count);
-}
-
-/**
-* printp - output a pointer data type to the standard output
-* @p: pointer to argument of type pointer
-* @count: existing total number of characters printed
-*
-* Return: cumulative total number of characters printed
-*/
-int printp(va_list p, int count)
-{
-	(void)p;
-	(void)count;
-	printf("printp");
-	return (count);
-}
-
-/**
-* printr - output a string to the standard output
-* @r: pointer to argument of type string literal
-* @count: existing total number of characters printed
-*
-* Return: cumulative total number of characters printed
-*/
-int printr(va_list r, int count)
-{
-	(void)r;
-	(void)count;
-	printf("printr");
 	return (count);
 }
